tegar/requirements.txt
flask
numpy
opencv-python
pandas
Flask-SQLAlchemy
scikit-learn
scikit-image
pillow